refactor(ui): extract SessionBrowser state hooks (fix lint)
