+++ packages/cli/src/ui/components/SessionBrowser/useSessionBrowserState.ts
@@ -1,0 +1,135 @@
+/**
+ * @license
+ * Copyright 2026 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { useState, useCallback, useMemo, useEffect, useRef } from 'react';
+import { useTerminalSize } from '../../hooks/useTerminalSize.js';
+import type { SessionInfo } from '../../../utils/sessionUtils.js';
+import type { SessionBrowserState } from '../SessionBrowser.js';
+import { sortSessions, filterSessions } from './utils.js';
+
+const SESSIONS_PER_PAGE = 20;
+
+/**
+ * Hook to manage all SessionBrowser state.
+ */
+export const useSessionBrowserState = (
+  initialSessions: SessionInfo[] = [],
+  initialLoading = true,
+  initialError: string | null = null,
+): SessionBrowserState => {
+  const { columns: terminalWidth } = useTerminalSize();
+  const [sessions, setSessions] = useState<SessionInfo[]>(initialSessions);
+  const [loading, setLoading] = useState(initialLoading);
+  const [error, setError] = useState<string | null>(initialError);
+  const [activeIndex, setActiveIndex] = useState(0);
+  const [scrollOffset, setScrollOffset] = useState(0);
+  const [sortOrder, setSortOrder] = useState<'date' | 'messages' | 'name'>(
+    'date',
+  );
+  const [sortReverse, setSortReverse] = useState(false);
+  const [searchQuery, setSearchQuery] = useState('');
+  const [isSearchMode, setIsSearchMode] = useState(false);
+  const [hasLoadedFullContent, setHasLoadedFullContent] = useState(false);
+  const loadingFullContentRef = useRef(false);
+
+  const filteredAndSortedSessions = useMemo(() => {
+    const filtered = filterSessions(sessions, searchQuery);
+    return sortSessions(filtered, sortOrder, sortReverse);
+  }, [sessions, searchQuery, sortOrder, sortReverse]);
+
+  // Reset full content flag when search is cleared
+  useEffect(() => {
+    if (!searchQuery) {
+      setHasLoadedFullContent(false);
+      loadingFullContentRef.current = false;
+    }
+  }, [searchQuery]);
+
+  const totalSessions = filteredAndSortedSessions.length;
+  const startIndex = scrollOffset;
+  const endIndex = Math.min(scrollOffset + SESSIONS_PER_PAGE, totalSessions);
+  const visibleSessions = filteredAndSortedSessions.slice(startIndex, endIndex);
+
+  const state: SessionBrowserState = {
+    sessions,
+    setSessions,
+    loading,
+    setLoading,
+    error,
+    setError,
+    activeIndex,
+    setActiveIndex,
+    scrollOffset,
+    setScrollOffset,
+    searchQuery,
+    setSearchQuery,
+    isSearchMode,
+    setIsSearchMode,
+    hasLoadedFullContent,
+    setHasLoadedFullContent,
+    sortOrder,
+    setSortOrder,
+    sortReverse,
+    setSortReverse,
+    terminalWidth,
+    filteredAndSortedSessions,
+    totalSessions,
+    startIndex,
+    endIndex,
+    visibleSessions,
+  };
+
+  return state;
+};
+
+/**
+ * Hook to handle selection movement.
+ */
+export const useMoveSelection = (state: SessionBrowserState) => {
+  const {
+    totalSessions,
+    activeIndex,
+    scrollOffset,
+    setActiveIndex,
+    setScrollOffset,
+  } = state;
+
+  return useCallback(
+    (delta: number) => {
+      const newIndex = Math.max(
+        0,
+        Math.min(totalSessions - 1, activeIndex + delta),
+      );
+      setActiveIndex(newIndex);
+
+      // Adjust scroll offset if needed
+      if (newIndex < scrollOffset) {
+        setScrollOffset(newIndex);
+      } else if (newIndex >= scrollOffset + SESSIONS_PER_PAGE) {
+        setScrollOffset(newIndex - SESSIONS_PER_PAGE + 1);
+      }
+    },
+    [totalSessions, activeIndex, scrollOffset, setActiveIndex, setScrollOffset],
+  );
+};
+
+/**
+ * Hook to handle sort order cycling.
+ */
+export const useCycleSortOrder = (state: SessionBrowserState) => {
+  const { sortOrder, setSortOrder } = state;
+
+  return useCallback(() => {
+    const orders: Array<'date' | 'messages' | 'name'> = [
+      'date',
+      'messages',
+      'name',
+    ];
+    const currentIndex = orders.indexOf(sortOrder);
+    const nextIndex = (currentIndex + 1) % orders.length;
+    setSortOrder(orders[nextIndex]);
+  }, [sortOrder, setSortOrder]);
+};
